finished day 3 part 2

+++ 2023/cpp Nov 2025/AdventOfCode/Day3.h
@@ -13,5 +13,5 @@ std::vector<std::string> getLinesFromFile();
 int PartOneReturnNumberIfValid(const std::vector<std::string>&lines, int row, int start, int stop);
 bool nextToSymbol(const std::vector<std::string>& lines, int row, int col);
 void Day3PartTwo();
-int returnProductIfAroundStar(std::vector<std::string>& lines, int row, int col);
+long long returnProductIfAroundStar(std::vector<std::string>& lines, int row, int col);
 bool isValidPoint(int row, int col, int  maxRows, int  maxCols);
